+++ module/JepRiaShowcase/App/client-react/src/App.tsx
@@ -2,14 +2,14 @@ import React, { Suspense, useEffect } from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Panel } from "@jfront/ui-core";
-import FeatureDetailPage from "./features/feature/pages/Detail";
-import FeatureCreatePage from "./features/feature/pages/Create";
-import FeatureSearchPage from "./features/feature/pages/Search";
-import FeatureListPage from "./features/feature/pages/List";
-import FeatureEditPage from "./features/feature/pages/Edit";
-import FeatureProcessListPage from "./features/feature-process/pages/List";
-import FeatureProcessDetailPage from "./features/feature-process/pages/Detail";
-import FeatureProcessCreatePage from "./features/feature-process/pages/Create";
+import FeatureDetailPage from "./features/feature/pages/FeatureDetailPage";
+import FeatureCreatePage from "./features/feature/pages/FeatureCreatePage";
+import FeatureSearchPage from "./features/feature/pages/FeatureSearchPage";
+import FeatureListPage from "./features/feature/pages/FeatureListPage";
+import FeatureEditPage from "./features/feature/pages/FeatureEditPage";
+import FeatureProcessListPage from "./features/feature-process/pages/FeatureProcessListPage";
+import FeatureProcessDetailPage from "./features/feature-process/pages/FeatureProcessDetailPage";
+import FeatureProcessCreatePage from "./features/feature-process/pages/FeatureProcessCreatePage";
 import AppTabPanel from "./features/tabpanel/AppTabPanel";
 import FeatureToolbar from "./features/feature/components/FeatureToolbar";
 import FeatureProcessToolbar from "./features/feature-process/components/FeatureProcessToolbar";
